feat: 실행 시간 측정용 어노테이션(@LogExecutionTime) 추가

+++ src/main/kotlin/com/ilgijjan/domain/diary/application/DiaryTaskProcessor.kt
@@ -1,5 +1,6 @@
 package com.ilgijjan.domain.diary.application
 
+import com.ilgijjan.common.annotation.LogExecutionTime
 import com.ilgijjan.domain.fcmtoken.application.FcmTokenDeleter
 import com.ilgijjan.domain.fcmtoken.application.FcmTokenReader
 import com.ilgijjan.integration.image.application.ImageGenerator
@@ -26,6 +27,7 @@ class DiaryTaskProcessor(
 ) {
     private val log = LoggerFactory.getLogger(javaClass)
 
+    @LogExecutionTime
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     fun process(diaryId: Long) {
         log.info("비동기 일기 생성 시작 - ID: $diaryId")
@@ -44,9 +46,11 @@ class DiaryTaskProcessor(
             val updateCommand = UpdateDiaryResultCommand.of(refinedText, imageUrl, musicResult)
             diaryUpdater.updateResult(diaryId, updateCommand)
 
-            val tokens = fcmTokenReader.findAllByUserId(diary.user.id!!).map { it.token }
-            val deadTokens = notificationSender.sendDiaryCompletion(tokens, diaryId)
-            fcmTokenDeleter.deleteByTokens(deadTokens)
+            if (diary.user.isNotificationEnabled) {
+                val tokens = fcmTokenReader.findAllByUserId(diary.user.id!!).map { it.token }
+                val deadTokens = notificationSender.sendDiaryCompletion(tokens, diaryId)
+                fcmTokenDeleter.deleteByTokens(deadTokens)
+            }
 
             log.info("비동기 일기 생성 완료 - ID: $diaryId")
         } catch (e: Exception) {
